feat: 현재 위치 이동 함수 커스텀훅으로 분리
