+++ src/app/(common)/_components/LocationButton.tsx
@@ -1,38 +1,10 @@
 "use client";
-import useLocationStore from "@/hooks/useLocationStore";
-import useMapStore from "@/hooks/useMapStore";
+import useMoveCurrentLocation from "@/hooks/useMoveCurrentLocation";
 import React from "react";
 import { MdOutlineMyLocation } from "react-icons/md";
 
 const LocationButton = () => {
-  const { changeCoordinates } = useLocationStore();
-  const { getMap } = useMapStore();
-  const getUserLocation = () => {
-    const map = getMap();
-    if (!map) return;
-    if (navigator.geolocation) {
-      navigator.geolocation.getCurrentPosition(
-        (position) => {
-          const lat = position.coords.latitude;
-          const lng = position.coords.longitude;
-          console.log(lat, lng, "hoho");
-          if (lat && lng) {
-            changeCoordinates(lat, lng);
-            const movePosition = new kakao.maps.LatLng(lat, lng);
-
-            map.setCenter(movePosition);
-          }
-        },
-
-        (error) => {
-          console.error(error);
-        },
-        { maximumAge: 5000 }
-      );
-    } else {
-      throw new Error("GPS 미지원");
-    }
-  };
+  const getUserLocation = useMoveCurrentLocation();
 
   return (
     <button
